Refactor SheetKind to SheetType and update imports accordingly
