+++ src/sheets/SheetData.ts
@@ -1,5 +1,5 @@
 import { SheetContent } from "./SheetContent";
-import { SheetType } from "./SheetKind";
+import { SheetType } from "./SheetType";
 import { SheetAlignment, SheetStyle } from "./SheetStyle";
 
 export type SheetData = SheetContent & SheetStyle & SheetAlignment & SheetType;
