+++ src/sheets/index.ts
@@ -2,7 +2,7 @@ export * from './SheetPosition';
 export * from './SheetRange';
 export * from './SheetSelector';
 
-export * from './SheetKind';
+export * from './SheetType';
 export * from './SheetStyle';
 export * from './SheetExpression';
 export * from './SheetRule';
